requirements.txt
psycopg2-binary==2.9.9
Django==5.1.2
djangorestframework==3.15.2
python-dotenv==1.0.1
gunicorn==23.0.0
pillow==11.0.0
requests